Add implementation of three for the price of two
I'm making the assumption here that the business wants us to apply the
discount equal to the least valuable of the three items. This will get
messy if we care about the order in which we consider sets of three, but
let's not worry about that yet.

Also remember to add all the new files this time...

Yes, the logic for calculating the minimum is grotesque. I'm sure
there's a better way.